move ping cmd + room emoji cmd desc
