+++ cogs/sysmgr.py
@@ -45,7 +45,6 @@ import ast
 import importlib
 import math
 import asyncio
-import time
 
 class Colors: # format: 0xHEXCODE
     greens_hair = 0xa19e78
@@ -1635,32 +1634,6 @@ class SysManager(commands.Cog, name=':wrench: System Manager'):
                 namematch = True
                 descmatch = True
                 match = 0
-
-    @commands.command()
-    async def ping(self, ctx):
-        """Measures bot latency."""
-        t = time.time()
-        msg = await ctx.send('Ping!')
-        diff = round((time.time() - t) * 1000, 1)
-        text = 'Pong! :ping_pong:'
-        if diff <= 300 and self.bot.latency <= 0.2:
-            embed = nextcord.Embed(title='Normal - all is well!',
-                                  description=f'Roundtrip: {diff}ms\nHeartbeat: {round(self.bot.latency * 1000, 1)}ms\n\nAll is working normally!',
-                                  color=0x00ff00)
-        elif diff <= 600 and self.bot.latency <= 0.5:
-            embed = nextcord.Embed(title='Fair - could be better.',
-                                  description=f'Roundtrip: {diff}ms\nHeartbeat: {round(self.bot.latency * 1000, 1)}ms\n\nNothing\'s wrong, but the latency could be better.',
-                                  color=0xffff00)
-        elif diff <= 2000 and self.bot.latency <= 1.0:
-            embed = nextcord.Embed(title='SLOW - __**oh no.**__',
-                                  description=f'Roundtrip: {diff}ms\nHeartbeat: {round(self.bot.latency * 1000, 1)}ms\n\nBot latency is higher than normal, messages may be slow to arrive.',
-                                  color=0xff0000)
-        else:
-            text = 'what'
-            embed = nextcord.Embed(title='**WAY TOO SLOW**',
-                                  description=f'Roundtrip: {diff}ms\nHeartbeat: {round(self.bot.latency * 1000, 1)}ms\n\nSomething is DEFINITELY WRONG here. Consider checking [Discord status](https://discordstatus.com) page.',
-                                  color=0xbb00ff)
-        await msg.edit(content=text, embed=embed)
 
 def setup(bot):
     bot.add_cog(SysManager(bot))
